[Storage] Move state store backup code to BackupHandler
Since we have the backup handler, we put all the backup related code to
the same place.
